requirements.txt
streamlit==1.25.0
transformers==4.25.1
PILLOW==9.5.0
torch==2.0.1
opencv-python-headless==4.7.0.72
moviepy==1.0.3
